Docker files and stuff

+++ Backend/Contracts/File/AppendCertificateRequest.cs
@@ -5,10 +5,10 @@ namespace Contracts.File
 {
     public class AppendCertificateRequest
     {
-        public double paddingLeft {  get; set; }
-        public double paddingTop { get; set; }
-        public double paddingRight { get; set; }        
-        public double paddingBottom { get; set; }
+        public float paddingLeft {  get; set; }
+        public float paddingTop { get; set; }
+        public float paddingRight { get; set; }        
+        public float paddingBottom { get; set; }
         public IFormFile? blank {  get; set; } = null;
     }
 }
